refactor-realtime-positions * Cleanup types used in use-realtime-vehicle-data and Map. Remove unused properties and move types into use-realtime-vehcile-data directory. * Remove filter-argument to useRealtimeVehicleData. Replace it with boundingBox argument. This was the only filter that was ever used. * Remove arguments to VehiclesQuery and VehicleUpdatesSubscription that was never passed
