+++ src/components/Map/RealtimeVehicleTag/RealtimeVehicleTag.tsx
@@ -3,7 +3,7 @@ import { Tooltip } from '@entur/tooltip'
 import { colors } from '@entur/tokens'
 import { LegMode, TransportMode } from '@entur/sdk'
 import { IconColorType } from '../../../types'
-import { RealtimeVehicle } from '../../../services/realtimeVehicles/types/realtimeVehicle'
+import { RealtimeVehicle } from '../../../logic/use-realtime-vehicle-data/types'
 import { useSettings } from '../../../settings/SettingsProvider'
 import { getIconColor } from '../../../utils/icon'
 import { TooltipContent } from './TooltipContent/TooltipContent'
