+++ src/logic/use-realtime-vehicle-data/VehiclesQuery.vehicles.graphql
@@ -1,21 +1,9 @@
 #import "./VehicleFragment.vehicles.graphql"
 
 query VehiclesQuery(
-    $codespaceId: String
-    $lineRef: String
-    $serviceJourneyId: String
-    $operatorRef: String
-    $mode: VehicleModeEnumeration
-    $monitored: Boolean
     $boundingBox: BoundingBox
 ) @api(name: vehicles) {
     vehicles(
-        codespaceId: $codespaceId
-        lineRef: $lineRef
-        serviceJourneyId: $serviceJourneyId
-        operatorRef: $operatorRef
-        mode: $mode
-        monitored: $monitored
         boundingBox: $boundingBox
     ) {
         ...VehicleFragment
