+++ src/logic/use-realtime-vehicle-data/types.ts
@@ -1,0 +1,21 @@
+type Line = {
+    lineRef: string
+    lineName?: string
+    publicCode?: string
+    pointsOnLink?: string
+}
+
+type RealtimeVehicle = {
+    vehicleRef: string
+    line: Line
+    mode: string
+    lastUpdated: string
+    lastUpdatedEpochSecond: number
+    location: {
+        latitude: number
+        longitude: number
+    }
+    active?: boolean
+}
+
+export { Line, RealtimeVehicle }
