--- src/services/realtimeVehicles/types/codespace.ts
@@ -1,3 +1,0 @@
-export type Codespace = {
-    codespaceId: string
-}
