--- src/services/realtimeVehicles/types/line.ts
@@ -1,6 +1,0 @@
-export type Line = {
-    lineRef: string
-    lineName?: string
-    publicCode?: string
-    pointsOnLink?: string
-}
