--- src/services/realtimeVehicles/types/operator.ts
@@ -1,3 +1,0 @@
-export type Operator = {
-    operatorRef: string
-}
