--- src/services/realtimeVehicles/types/serviceJourney.ts
@@ -1,3 +1,0 @@
-export type ServiceJourney = {
-    serviceJourneyId: string
-}
